fix(agent): harden lint reflection and fix audit findings
