+++ src/lifecycle-evaluators.ts
@@ -115,17 +115,22 @@ export const lintEvaluator: Evaluator = {
   id: "lint",
   evaluate(ctx) {
     if (ctx.mode !== "work" || !ctx.workspace) return { type: "done" };
+    if (!ctx.policy.lintCommand) return { type: "done" };
     const paths = writePathsForCurrentTask(ctx);
     if (paths.length === 0) return { type: "done" };
-    const result = lintFiles(ctx.workspace, paths);
+    const result = lintFiles(ctx.workspace, paths, ctx.policy.lintCommand);
     if (!result.hasErrors) return { type: "done" };
     ctx.debug("lifecycle.eval.lint", { files: paths.length });
     return {
       type: "regenerate",
-      prompt:
-        `${ctx.agentInput}\n\n` +
-        `Lint errors detected in files you edited:\n${result.output}\n\n` +
+      prompt: [
+        ctx.agentInput,
+        "",
+        "Lint errors detected in files you edited:",
+        result.output,
+        "",
         "If the project has an auto-fix command, run it first. Otherwise fix the errors manually, then stop.",
+      ].join("\n"),
     };
   },
 };
